# openapi.json
{"openapi":"3.1.0","info":{"title":"2022 Durban Property Valuations API","summary":"An API that for property valuations in Durban in 2022.","description":"This API returns data on the valuations of Full and Sectional title properties in Durban in 2022. The data returned is scraped from https://valuation2022.durban.gov.za/.","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/get_properties/property_type/{property_type}":{"get":{"summary":"Get Data","operationId":"get_data_get_properties_property_type__property_type__get","parameters":[{"name":"property_type","in":"path","required":true,"schema":{"enum":["FULL","SECT"],"type":"string","title":"Property Type"}},{"name":"rate_number","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rate Number"}},{"name":"legal_description","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Legal Description"}},{"name":"address","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"}},{"name":"first_owner","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Owner"}},{"name":"use_code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Use Code"}},{"name":"rating_category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rating Category"}},{"name":"market_value","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Market Value"}},{"name":"registered_extent","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Registered Extent"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}